# leveldb: implement concurrent compaction