This PR runs the isolated compaction concurrently to raise the system performance.