@rjl493456442 2020-08-15
A few more information after running the write benchmarking.

### Benchmarking setting

- Key size 100b
- Value size 100b
- Total value size 10GB
- Write buffer 100MB

### Master

- Database statistic

```
{38290 15m11.250540464s false 0 0 96677349193 66976108702 0 4 [320361511 2492915337 12383826915] 
[3 1481 6045] [0 29072592687 37897539069] [15270499841 29047698903 37849036775] [1m56.873115244s 8m39.370564199s 11m49.064456104s] 143 16 5618 0}
```

- Performance statistic

```
-- batch-100kb.2020-08-06-18:35:19 (17476 events) total time: 1264.8400s
 total size: 10737254400 bytes
  mean mb/s: 8.096 (+- 12.743)
```

### PR 

- Database statistic 

```
{42624 8m53.29404969s false 0 0 119958158300 90251417268 0 33 [427150354 1210462062 1680973110 11882306866] 
[4 623 975 6111] [0 17962660959 27789895644 44474303957] [15270499841 17954901062 27777240464 44425111585] 
[2m33.999790202s 10m35.894896992s 34m55.40700714s 1h0m21.973738336s] 143 20 10243 0}
```

- Performance statistic

```
-- batch-100kb.2020-08-06-19:04:43 (17476 events) total time: 981.0566s
 total size: 10737254400 bytes
  mean mb/s: 10.438 (+- 12.787)
```