@rjl493456442 2020-08-15
### Benchmarking setting 2

- Key size 100b
- Value size 100b
- Total value size **20GB**
- Write buffer 100MB

### Master

- Database statistic

```
{74923 51m55.158298393s false 0 0 217969445106 158587878647 0 6 [1281435560 2343980577 12772278361 13972973946] [12 1380 6216 6911] [0 57512431061 101038735425 25313162] [30541031041 57460130601 100920691029 25295421] [3m57.212215565s 21m38.221922474s 39m40.466237826s 598.812645ms] 286 27 11667 0}
```

- Performance statistic

```
-- batch-100kb.2020-08-06-20:48:44 (34952 events) total time: 3845.0401s
 total size: 21474508800 bytes
  mean mb/s: 5.326 (+- 11.064)
```

### PR 

- Database statistic 

```
{79857 32m41.074744105s false 0 0 285748221868 226365733290 0 3 [1067864748 855142281 2348273262 16024399610 10071344883] [10 450 1421 8411 5194] [0 35578820817 56449218779 134285576309 52117385] [30541031041 35561629964 56415397892 134162611693 52087484] [5m14.606409304s 33m38.251109674s 1h22m49.8935168s 3h28m28.534348668s 3.998871177s] 286 31 24112 0}
```

- Performance statistic

```
-- batch-100kb.2020-08-06-19:58:29 (34952 events) total time: 2723.2274s
 total size: 21474508800 bytes
  mean mb/s: 7.520 (+- 12.231)
```