@syndtr 2020-08-16
Hi @rjl493456442,

I did quick stress test and this seems to cause some data inconsistency.

```
2020/08/16 19:34:02 FATAL: [02] READER #31.24 K474 snap.Get: leveldb: not found
k1: 0201dde13fa10e61f346b0b2dcb1b286a1735eb555880100007bc4c2780201dde13fa10e61f346b0b2dcb1b286a1735eb555880100007bc4c2784fd8259bb3
 -> k2: 0202409476998cd27fb4ed61d9e9099aae7b2916c71f0000003a2d00990202409476998cd27fb4ed61d9e9099aae7b2916c71f0000003a2d00995347a528f7
2020/08/16 19:34:02 [02] READER #31.24 DONE Snap=leveldb.Snapshot{6871984} Alive=1 IterLatency=1.627µs GetLatency=22.355µs
2020/08/16 19:34:02 [05] READER #24.367 DONE Snap=leveldb.Snapshot{6422524} Alive=1 IterLatency=1.251µs GetLatency=17.331µs
2020/08/16 19:34:02 [03] READER #25.341 DONE Snap=leveldb.Snapshot{6288541} Alive=1 IterLatency=1.273µs GetLatency=20.634µs
2020/08/16 19:34:02 [01] READER #26.280 DONE Snap=leveldb.Snapshot{6418609} Alive=1 IterLatency=1.204µs GetLatency=18.291µs
2020/08/16 19:34:02 [06] READER #15.2 DONE Snap=leveldb.Snapshot{6808439} Alive=1 IterLatency=2.619µs GetLatency=43.109µs
2020/08/16 19:34:02 [02] READER #30.667 DONE Snap=leveldb.Snapshot{6423586} Alive=0 IterLatency=1.249µs GetLatency=18.231µs
2020/08/16 19:34:02 [03] READER #26.85 DONE Snap=leveldb.Snapshot{6736953} Alive=0 IterLatency=1.511µs GetLatency=31.515µs
2020/08/16 19:34:02 [05] READER #25.36 DONE Snap=leveldb.Snapshot{6870922} Alive=0 IterLatency=1.535µs GetLatency=9.862µs
2020/08/16 19:34:02 [01] READER #27.20 DONE Snap=leveldb.Snapshot{6867009} Alive=0 IterLatency=1.517µs GetLatency=24.941µs
2020/08/16 19:34:02 [04] READER #19.3 DONE Snap=leveldb.Snapshot{6864345} Alive=1 IterLatency=1.778µs GetLatency=27.333µs
2020/08/16 19:34:02 [04] READER #18.39 DONE Snap=leveldb.Snapshot{6414921} Alive=0 IterLatency=1.292µs GetLatency=18.948µs
2020/08/16 19:34:02 [04] SCANNER #16 Deleted=9562 Time=4.125932536s
2020/08/16 19:34:02 [04] SCANNER DONE #17
2020/08/16 19:34:02 [03] SCANNER #17 Deleted=1236 Time=4.13560786s
2020/08/16 19:34:02 [03] SCANNER DONE #18
2020/08/16 19:34:02 > Write using transaction
2020/08/16 19:34:02 [06] READER #14.11 DONE Snap=leveldb.Snapshot{6360027} Alive=0 IterLatency=1.673µs GetLatency=21.88µs
2020/08/16 19:34:02 [00] READER #13.3 DONE Snap=leveldb.Snapshot{6275273} Alive=1 IterLatency=1.707µs GetLatency=33.257µs
2020/08/16 19:34:02 [03] WRITER DONE #28
2020/08/16 19:34:02 [03] READER #27.0 DONE Snap=leveldb.Snapshot{6985340} Alive=0 IterLatency=0s GetLatency=0s
2020/08/16 19:34:03 [00] READER #14.2 DONE Snap=leveldb.Snapshot{6723692} Alive=0 IterLatency=1.561µs GetLatency=21.968µs
panic: negative ref: 3933

goroutine 8 [running]:
github.com/syndtr/goleveldb/leveldb.(*session).refLoop.func1(0xf5d, 0xffffffffffffffff, 0xf61)
	/home/syndtr/pg/githubproj/goleveldb/leveldb/session_util.go:89 +0x160
github.com/syndtr/goleveldb/leveldb.(*session).refLoop.func3(0xc005dadb80)
	/home/syndtr/pg/githubproj/goleveldb/leveldb/session_util.go:107 +0xe8
github.com/syndtr/goleveldb/leveldb.(*session).refLoop.func4()
	/home/syndtr/pg/githubproj/goleveldb/leveldb/session_util.go:176 +0x3a9
github.com/syndtr/goleveldb/leveldb.(*session).refLoop(0xc0000d26c0)
	/home/syndtr/pg/githubproj/goleveldb/leveldb/session_util.go:187 +0x44b
created by github.com/syndtr/goleveldb/leveldb.newSession
	/home/syndtr/pg/githubproj/goleveldb/leveldb/session.go:95 +0x2b4
exit status 2
```

I din't dig into the code, but if you do multi-level concurrent compaction you need to take extra precaution.

---

But, I've decided not to merge this or any non-trivial PR, since I most likely to don't have time and energy to test and validate those. 
I'm very sorry about that. Feel free to fork the project though.
Thanks for the effort Gary, appreciate it :)