@qianbin 2020-08-28
Hi Gary, I'm also working on concurrent compaction. Could you please share your bench code ?